services/stock-fetching-service/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
beautifulsoup4==4.13.3
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.2.2
fastapi==0.115.8
frozendict==2.4.6
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
kafka-python==2.0.2
kombu==5.4.2
loguru==0.7.3
lxml==5.3.1
multitasking==0.0.11
numpy==2.2.2
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.6
prompt_toolkit==3.0.50
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.2
yfinance==0.2.52
